# improve readBytesOrString especially for readAll