--- a/test/io/ferguson/utf8/readstring-stdin-binary-utf8.chpl
+++ b/test/io/ferguson/utf8/readstring-stdin-binary-utf8.chpl
@@ -7,7 +7,7 @@ var style = stdin._styleInternal();
 style.binary = 1;
 stdin._set_styleInternal(style);
 
-// read 10 bytes
+// read 10 codepoints
 stdin.readString(str, 10);
 writeln("Read\n", str);