--- a/test/io/ferguson/utf8/readstring-stdin-text-utf8.good
+++ b/test/io/ferguson/utf8/readstring-stdin-text-utf8.good
@@ -1,8 +1,8 @@
 Read
 This ☃
-.
+.is
 Read
-is
+
 . a
 .  test.