… runtime/include/qio/qio.h
… test/io/ferguson/utf8/readstring-stdin-text-utf8.chpl
… test/io/ferguson/utf8/various-reads-invalid-utf8.chpl
… test/io/ferguson/utf8/various-reads-invalid-utf8.good
… test/library/standard/IO/readAll/readStringInvalid.chpl
… test/library/standard/IO/readAll/readStringInvalid.good
… test/library/standard/IO/readStringAndBytes/checkAllocationSize.chpl
… test/library/standard/IO/readStringAndBytes/checkAllocationSize.good
… test/library/standard/IO/readStringAndBytes/moby.txt
… test/library/standard/IO/readStringAndBytes/readBytes.chpl
… test/library/standard/IO/readStringAndBytes/readString.chpl
… test/library/standard/IO/readStringAndBytes/readStringUnicode.chpl
… test/library/standard/IO/readStringAndBytes/readStringUnicode.good
… test/library/standard/IO/readStringAndBytes/snowmen.txt