This PR improves the performance of I/O routines that were calling `readBytesOrString` : `readString`, `readBytes`, `readAll`, and `readBinary(string|bytes)`. This PR is motivated by wanting to improve the performance of the regexredux2 and regexredux3 CLBG benchmarks which use `readAll`. We identified that the `readAll` call was slower than similar code in C++.

The approach here is to factor `readBytesOrString` into `readBytesImpl` and `readStringImpl`.

`readBytesImpl` uses repeated calls to `qio_channel_read` while doubling the size of the buffer as needed to store more data in it. It uses the initial file size (which is already computed in the I/O library and just needed a function to get at it) to compute an initial buffer size. This avoids several kinds of overhead and forms an alternative implementation of #19270.

`readStringImpl` is similar, but it reads a codepoint at a time, in order to preserve the end offset when invalid UTF-8 is encountered (see PR #24807).

This PR also resolves #24781 by adjusting the documentation for `string.createAdoptingBuffer` to clarify that the buffer is not freed when an error is thrown.

This PR removes some special case code that was making `readString` use `maxSize` in bytes, rather than in codepoints, when the undocumented binary mode. This makes `readString` more consistent.

This PR changes the behavior on error for the `readString`, `readBinary(string/bytes)`, and `readAll` functions. It changes them so that, on an error such as invalid UTF-8, the channel position is left near the error. The alternative behavior of returning to the start of the read on such an error would require extra memory / buffering and reduce performance. Anyway, the offset after trying to decode invalid UTF-8 is now tested in `test/io/ferguson/utf8/various-reads-invalid-utf8.chpl`.

Lastly, this PR updates the documentation for a number of `read*` functions in order to clarify where the channel position is left on an error.

How does it affect performance? These measurements were collected on my PC with `start_test -performance test/studies/shootout/submitted/regexredux*.chpl --numtrials 4` (and note that `main` here includes #24757). Please note that the performance gain for `readAll` itself will be a larger percentage (since it is not most of these benchmarks).

| Benchmark    | main   | this PR | speedup    |
| ------------ | ------ | ------- | ---------- |
| regexredux2  | 1.55 s | 1.45 s  |  7% faster |
| regexredux3  | 1.29 s | 1.17 s  | 10% faster |

Reviewed by @vasslitvinov and @jeremiah-corrado - thanks!

- [x] full comm=none testing